Clean up splint warnings: check more return values, use more logical types etc

+++ utilities.h
@@ -5,7 +5,7 @@ typedef struct { float x, y, z; } vertex;
 typedef struct { float r, g, b; } colour;
 
 typedef struct{
-	int sides;
+	unsigned short sides;
 	long int *verts;
 } face;
 
@@ -14,14 +14,14 @@ void read_vertex_data(FILE * infile
 	, vertex *vertices
 	, vector *normals
 	, colour *colours
-	, long int numverts
+	, unsigned long numverts
 	, int has_normals
 	, int has_colours );
 
 
 void read_face_data( FILE * infile
 	, face* faces
-	, long int numfaces );
+	, unsigned long numfaces );
 
 
 void write_off_file( FILE * outfile
@@ -29,9 +29,9 @@ void write_off_file( FILE * outfile
 	, vector * normals
 	, face * faces
 	, colour * colours
-	, long int numverts
-	, long int numfaces
-	, long int numedges
+	, unsigned long numverts
+	, unsigned long numfaces
+	, unsigned long numedges
 	, int has_normals
 	, int has_colour );
